+++ pkg/issuer/acme/dns/rfc2136/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
     deps = [
         "//pkg/acme/webhook/apis/acme/v1alpha1:go_default_library",
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
+        "//pkg/apis/certmanager/validation/util:go_default_library",
         "//vendor/github.com/miekg/dns:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
         "//vendor/k8s.io/client-go/informers:go_default_library",
